@@ -42,7 +42,7 @@ class Kohana_ConfigTest extends Unittest_TestCase
 	public function test_attach_adds_reader_and_returns_this()
 	{
 		$config = new Config;
-		$reader = $this->getMock('Kohana_Config_Reader');
+		$reader = $this->createMock('Kohana_Config_Reader');
 
 		$this->assertSame($config, $config->attach($reader));
 
@@ -60,8 +60,8 @@ class Kohana_ConfigTest extends Unittest_TestCase
 	{
 		$config  = new Config;
 
-		$reader1 = $this->getMock('Kohana_Config_Reader');
-		$reader2 = $this->getMock('Kohana_Config_Reader');
+		$reader1 = $this->createMock('Kohana_Config_Reader');
+		$reader2 = $this->createMock('Kohana_Config_Reader');
 
 		$config->attach($reader1);
 		$config->attach($reader2);
@@ -89,8 +89,8 @@ class Kohana_ConfigTest extends Unittest_TestCase
 	public function test_attach_can_add_reader_to_end_of_queue()
 	{
 		$config  = new Config;
-		$reader1 = $this->getMock('Kohana_Config_Reader');
-		$reader2 = $this->getMock('Kohana_Config_Reader');
+		$reader1 = $this->createMock('Kohana_Config_Reader');
+		$reader2 = $this->createMock('Kohana_Config_Reader');
 
 		$config->attach($reader1);
 		$config->attach($reader2, FALSE);
@@ -112,8 +112,10 @@ class Kohana_ConfigTest extends Unittest_TestCase
 		// that has already been used then it just re-uses the first's name
 
 		// To get around this we have to specify a totally random name for the second mock object
-		$reader1 = $this->getMock('Kohana_Config_Reader');
-		$reader2 = $this->getMock('Kohana_Config_Reader', array(), array(), 'MY_AWESOME_READER');
+		$reader1 = $this->createMock('Kohana_Config_Reader');
+		$reader2 = $this->getMockBuilder('Kohana_Config_Reader')
+            ->setConstructorArgs([ array(), array(), 'MY_AWESOME_READER'])
+            ->getMock();
 
 		$config->attach($reader1);
 		$config->attach($reader2);
@@ -137,7 +139,7 @@ class Kohana_ConfigTest extends Unittest_TestCase
 	public function test_detach_returns_this_even_when_reader_dnx()
 	{
 		$config = new Config;
-		$reader = $this->getMock('Kohana_Config_Reader');
+		$reader = $this->createMock('Kohana_Config_Reader');
 
 		$this->assertSame($config, $config->detach($reader));
 	}
@@ -153,7 +155,7 @@ class Kohana_ConfigTest extends Unittest_TestCase
 	{
 		$config = new Config;
 
-		$reader = $this->getMock('Kohana_Config_Reader', array('load'));
+		$reader = $this->getMockBuilder('Kohana_Config_Reader')->setConstructorArgs([array('load')])->getMock();
 
 		$reader
 			->expects($this->once())
@@ -178,7 +180,7 @@ class Kohana_ConfigTest extends Unittest_TestCase
 	{
 		$config = new Config;
 
-		$reader = $this->getMock('Kohana_Config_Reader', array('load'));
+		$reader = $this->getMockBuilder('Kohana_Config_Reader')->setConstructorArgs([array('load')])->getMock();
 
 		$reader
 			->expects($this->once())
@@ -240,7 +242,7 @@ class Kohana_ConfigTest extends Unittest_TestCase
 	{
 		$config = new Kohana_Config;
 
-		$reader = $this->getMock('Kohana_Config_Reader');
+		$reader = $this->createMock('Kohana_Config_Reader');
 
 		$config->attach($reader);
 
@@ -258,9 +260,9 @@ class Kohana_ConfigTest extends Unittest_TestCase
 	{
 		$config = new Kohana_Config;
 
-		$reader1 = $this->getMock('Kohana_Config_Reader');
-		$writer1 = $this->getMock('Kohana_Config_Writer', array('write'));
-		$writer2 = $this->getMock('Kohana_Config_Writer', array('write'));
+		$reader1 = $this->createMock('Kohana_Config_Reader');
+		$writer1 = $this->getMockBuilder('Kohana_Config_Writer')->setConstructorArgs([array('write')])->getMock();
+		$writer2 = $this->getMockBuilder('Kohana_Config_Writer')->setConstructorArgs([array('write')])->getMock();
 
 		$writer1
 			->expects($this->once())
@@ -288,8 +290,12 @@ class Kohana_ConfigTest extends Unittest_TestCase
 	{
 		$group_name =  'lolumns';
 
-		$reader1 = $this->getMock('Kohana_Config_Reader', array('load'), array(), 'Unittest_Config_Reader_1');
-		$reader2 = $this->getMock('Kohana_Config_Reader', array('load'), array(), 'Unittest_Config_Reader_2');
+		$reader1 = $this->getMockBuilder('Kohana_Config_Reader')
+            ->setConstructorArgs([ array('load'), array(), 'Unittest_Config_Reader_1'])
+            ->getMock();
+		$reader2 = $this->getMockBuilder('Kohana_Config_Reader')
+            ->setConstructorArgs([ array('load'), array(), 'Unittest_Config_Reader_2'])
+            ->getMock();
 
 		$reader1
 			->expects($this->once())
@@ -333,7 +339,7 @@ class Kohana_ConfigTest extends Unittest_TestCase
 	 */
 	public function test_load_reuses_config_groups()
 	{
-		$reader = $this->getMock('Kohana_Config_Reader', array('load'));
+		$reader = $this->getMockBuilder('Kohana_Config_Reader')->setConstructorArgs([array('load')])->getMock();
 		$reader
 			->expects($this->once())
 			->method('load')
@@ -360,8 +366,8 @@ class Kohana_ConfigTest extends Unittest_TestCase
 	{
 		$config = new Kohana_Config;
 
-		$reader1 = $this->getMock('Kohana_Config_Reader', array('load'));
-		$reader2 = $this->getMock('Kohana_Config_Reader', array('load'));
+		$reader1 = $this->getMockBuilder('Kohana_Config_Reader')->setConstructorArgs([array('load')])->getMock();
+		$reader2 = $this->getMockBuilder('Kohana_Config_Reader')->setConstructorArgs([array('load')])->getMock();
 
 		$reader1
 			->expects($this->once())
@@ -375,8 +381,8 @@ class Kohana_ConfigTest extends Unittest_TestCase
 			->with('something')
 			->will($this->returnValue(array('kohana' => 'good')));
 
-		$writer1 = $this->getMock('Kohana_Config_Writer', array('write'));
-		$writer2 = $this->getMock('Kohana_Config_Writer', array('write'));
+		$writer1 = $this->getMockBuilder('Kohana_Config_Writer')->setConstructorArgs([array('write')])->getMock();
+		$writer2 = $this->getMockBuilder('Kohana_Config_Writer')->setConstructorArgs([array('write')])->getMock();
 
 		// Due to crazy limitations in phpunit's mocking engine we have to be fairly
 		// liberal here as to what order we receive the config items
