@@ -175,13 +175,7 @@ class Kohana_RouteTest extends Unittest_TestCase
 	public function test_constructor_returns_if_uri_is_null()
 	{
 		// We use a mock object to make sure that the route wasn't recompiled
-		$route = $this->getMock('Route', array('_compile'), array(), '', FALSE);
-
-		$route
-			->expects($this->never())
-			->method('_compile');
-
-		$route->__construct(NULL,NULL);
+		$route = new MockRoute(NULL, NULL);
 
 		$this->assertAttributeSame('', '_uri', $route);
 		$this->assertAttributeSame(array(), '_regex', $route);
@@ -521,7 +515,7 @@ class Kohana_RouteTest extends Unittest_TestCase
 	 */
 	public function test_defaults_throws_exception_when_setting_fqcn_and_directory()
 	{
-		$this->setExpectedException('Kohana_Exception', 'Route directory should not be set when the controller is a FQCN.');
+		$this->expectException('Kohana_Exception', 'Route directory should not be set when the controller is a FQCN.');
 
 		$route = new Route('(<controller>(/<action>(/<id>)))');
 		$route->defaults(array(
@@ -656,7 +650,7 @@ class Kohana_RouteTest extends Unittest_TestCase
 	{
 		$route = new Route($uri, $regex);
 
-		$this->setExpectedException('Kohana_Exception', 'controller');
+		$this->expectException('Kohana_Exception', 'controller');
 		$route->uri($uri_array);
 	}
 
@@ -879,5 +873,14 @@ class Kohana_RouteTest extends Unittest_TestCase
 	{
 	    return \Request::with(['uri' => $uri, 'method' => \Request::GET]);
 	}
+
+}
+
+class MockRoute extends Route
+{
+    public static function compile($uri, array $regex = NULL)
+    {
+        throw new \BadMethodCallException('This should not be being called');
+    }
 
 }
