@@ -610,7 +610,7 @@ class Kohana_ValidationTest extends Unittest_TestCase
 	public function test_offsetSet_throws_exception()
 	// @codingStandardsIgnoreEnd
 	{
-		$this->setExpectedException('Kohana_Exception');
+		$this->expectException('Kohana_Exception');
 
 		$validation = Validation::factory(array());
 
@@ -639,7 +639,7 @@ class Kohana_ValidationTest extends Unittest_TestCase
 	public function test_offsetUnset()
 	// @codingStandardsIgnoreEnd
 	{
-		$this->setExpectedException('Kohana_Exception');
+		$this->expectException('Kohana_Exception');
 
 		$validation = Validation::factory(array(
 			'one' => 'Hello, World!',
